+++ NEWS.md
@@ -6,7 +6,7 @@
 * Added function *find_countrycol()* to automate search of columns containing country names in data frames
 * Added function *find_timecol()* to automate search of date and/or year columns in data frames
 * Added function *find_keycol()* to automate search of table keys in data frames
-* Added functions *which_min()*, *which_max()* and *which_mode()*
+* Added functions *which_min()*, *which_max()* and *which_mode()* returning all the positions of vector's minimum, maximum and mode values.
 * Speed improvements for the functions *country_name()* and *match_table()*
 * Bug fixes
 * Added vignette on data structure